# api/openapi.yaml
openapi: 3.0.3
info:
  title: GoLedger API
  description: Double-entry ledger system with account management, transfers, and holds
  version: 1.0.0
  contact:
    name: API Support
    email: support@goledger.io
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.goledger.io/v1
    description: Production server
...
paths:
...
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Returns current authenticated user information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Accounts
...
  /accounts/{id}/transfers:
    get:
      tags: [Transfers]
      summary: List transfers for an account
      description: >
        Pass `cursor` (the ID of the last transfer from the previous page) for
        keyset pagination - stable under concurrent writes, and the
        recommended approach. Omit `cursor` entirely to fall back to legacy
        `offset` pagination.
      operationId: listTransfersByAccount
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: cursor
          in: query
          description: ID of the last transfer seen; triggers keyset pagination when present (even if empty, to start from the most recent transfer).
          schema:
            type: string
      responses:
        '200':
          description: List of transfers (cursor-paginated shape shown; offset mode returns a bare array)
          content:
            application/json:
              schema:
                type: object
                properties:
                  transfers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transfer'
                  next_cursor:
                    type: string
                    description: Pass as `cursor` to fetch the next page; absent/empty on the last page.
...
  /transfers:
    post:
      tags: [Transfers]
      summary: Create transfer
      description: Create a new transfer between accounts
      operationId: createTransfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '201':
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '412':
          description: Precondition failed (insufficient funds, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transfers/batch:
    post:
      tags: [Transfers]
      summary: Batch transfer
      description: Create multiple transfers atomically in a single transaction.
      operationId: createBatchTransfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transfers]
              properties:
                event_at:
                  type: string
                  format: date-time
                metadata:
                  type: object
                  additionalProperties: true
                transfers:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '201':
          description: Transfers created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '412':
          description: Precondition failed (insufficient funds, currency mismatch, etc.) - none of the batch is applied
  /transfers/{id}:
    get:
      tags: [Transfers]
      summary: Get transfer
      description: Get transfer by ID
      operationId: getTransfer
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /transfers/{id}/reverse:
    post:
      tags: [Transfers]
      summary: Reverse transfer
      description: Create a reversal transfer
      operationId: reverseTransfer
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Reversal transfer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '412':
          description: Transfer already reversed or cannot be reversed

  # Holds
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of results (max 1000)
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    UserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, operator, viewer]
...
    Transfer:
      type: object
      properties:
        id:
          type: string
        from_account_id:
          type: string
        to_account_id:
          type: string
        amount:
          type: string
          description: Transfer amount (decimal string)
        created_at:
          type: string
          format: date-time
        event_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        reversed_transfer_id:
          type: string
          nullable: true
    CreateTransferRequest:
      type: object
      required: [from_account_id, to_account_id, amount]
      properties:
        from_account_id:
          type: string
        to_account_id:
          type: string
        amount:
          type: string
          pattern: '^\d+(\.\d+)?$'
          example: "100.00"
        metadata:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'