# internal/infrastructure/postgres/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package generated

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID                   string             `json:"id"`
	Name                 string             `json:"name"`
	Currency             string             `json:"currency"`
	Balance              pgtype.Numeric     `json:"balance"`
	Version              int64              `json:"version"`
	AllowNegativeBalance bool               `json:"allow_negative_balance"`
	AllowPositiveBalance bool               `json:"allow_positive_balance"`
	CreatedAt            pgtype.Timestamptz `json:"created_at"`
	UpdatedAt            pgtype.Timestamptz `json:"updated_at"`
	EncumberedBalance    pgtype.Numeric     `json:"encumbered_balance"`
}

type AuditLog struct {
	ID           string             `json:"id"`
	UserID       string             `json:"user_id"`
	Action       string             `json:"action"`
	ResourceType string             `json:"resource_type"`
	ResourceID   string             `json:"resource_id"`
	IpAddress    *string            `json:"ip_address"`
	UserAgent    *string            `json:"user_agent"`
	RequestID    *string            `json:"request_id"`
	BeforeState  []byte             `json:"before_state"`
	AfterState   []byte             `json:"after_state"`
	Status       string             `json:"status"`
	ErrorMessage *string            `json:"error_message"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	PrevHash     *string            `json:"prev_hash"`
	Hash         *string            `json:"hash"`
	ChainSeq     *int64             `json:"chain_seq"`
}

type AuditLogChainState struct {
	ID       bool   `json:"id"`
	LastHash string `json:"last_hash"`
	LastSeq  int64  `json:"last_seq"`
}

type AuditLogsDefault struct {
	ID           string             `json:"id"`
	UserID       string             `json:"user_id"`
	Action       string             `json:"action"`
	ResourceType string             `json:"resource_type"`
	ResourceID   string             `json:"resource_id"`
	IpAddress    *string            `json:"ip_address"`
	UserAgent    *string            `json:"user_agent"`
	RequestID    *string            `json:"request_id"`
	BeforeState  []byte             `json:"before_state"`
	AfterState   []byte             `json:"after_state"`
	Status       string             `json:"status"`
	ErrorMessage *string            `json:"error_message"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type Entry struct {
	ID                     string             `json:"id"`
	AccountID              string             `json:"account_id"`
	TransferID             string             `json:"transfer_id"`
	Amount                 pgtype.Numeric     `json:"amount"`
	AccountPreviousBalance pgtype.Numeric     `json:"account_previous_balance"`
	AccountCurrentBalance  pgtype.Numeric     `json:"account_current_balance"`
	AccountVersion         int64              `json:"account_version"`
	CreatedAt              pgtype.Timestamptz `json:"created_at"`
}

type Hold struct {
	ID        string             `json:"id"`
	AccountID string             `json:"account_id"`
	Amount    pgtype.Numeric     `json:"amount"`
	Status    string             `json:"status"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	Metadata  []byte             `json:"metadata"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type OutboxEvent struct {
	ID                string             `json:"id"`
	AggregateID       string             `json:"aggregate_id"`
	AggregateType     string             `json:"aggregate_type"`
	EventType         string             `json:"event_type"`
	Payload           []byte             `json:"payload"`
	CreatedAt         pgtype.Timestamptz `json:"created_at"`
	PublishedAt       pgtype.Timestamptz `json:"published_at"`
	Published         bool               `json:"published"`
	EventVersion      int32              `json:"event_version"`
	AggregateSequence int64              `json:"aggregate_sequence"`
	Attempts          int32              `json:"attempts"`
	LastError         *string            `json:"last_error"`
	DeadLetteredAt    pgtype.Timestamptz `json:"dead_lettered_at"`
}

type Transfer struct {
	ID                 string             `json:"id"`
	FromAccountID      string             `json:"from_account_id"`
	ToAccountID        string             `json:"to_account_id"`
	Amount             pgtype.Numeric     `json:"amount"`
	CreatedAt          pgtype.Timestamptz `json:"created_at"`
	EventAt            pgtype.Timestamptz `json:"event_at"`
	Metadata           []byte             `json:"metadata"`
	ReversedTransferID *string            `json:"reversed_transfer_id"`
}

type User struct {
	ID             string             `json:"id"`
	Email          string             `json:"email"`
	Name           string             `json:"name"`
	HashedPassword string             `json:"hashed_password"`
	Role           string             `json:"role"`
	Active         bool               `json:"active"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}